+++ src/Trait/ListenerTrait.php
@@ -20,6 +20,8 @@ trait ListenerTrait
     /* @todo: Deprecate and remove in v4 */
     public function onConnect(Closure $closure): self
     {
+        $msg = 'onConnect() is deprecated and will be removed in v4. Use onHandshake() instead.';
+        trigger_error($msg, E_USER_DEPRECATED);
         $this->listeners['connect'] = $closure;
         return $this;
     }