+++ docs/Changelog.md
@@ -12,6 +12,7 @@
  * `onHandshake(...)` listener (will deprecate `onConnect(...)`) (@sirn-se)
  * Server `setMaxConnections(int|null)` (@sirn-se)
  * Server `shutdown()` orderly close server (@sirn-se)
+ * `phrity/net-uri v2.1` for URI decoding (@sirn-se)
 
 ### `3.0.0`